requirements.txt
certifi==2018.4.16
chardet==3.0.4
confusable-homoglyphs==3.0.0
diff-match-patch==20121119
Django==2.0.5
django-countries==5.2
django-import-export==1.0.0
django-registration-redux==2.4
djangorestframework==3.8.2
docopt==0.6.2
et-xmlfile==1.0.1
idna==2.6
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.3
pipreqs==0.4.9
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.4
PyYAML==3.12
requests==2.18.4
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
yarg==0.1.9
